+++ Entity/Models/Reports/Project.cs
@@ -10,7 +10,7 @@ namespace Entity.Models
         public uint ReportId { get; set; }
         public ICollection<Report> Reports { get; set; }
 
-        public virtual uint UserId { get; set; }
+        public virtual long UserId { get; set; }
         public virtual User User { get; set; }
     }
 }
